# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Navigator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface CareerRecommendBody {
  interests: string;
  skills: string;
  skillLevel?: string;
  goals: string;
  workLifePreference?: string;
  field?: string;
}

export interface CareerOption {
  name: string;
  type: string;
  matchPercent: number;
  salary: string;
  demand: string;
  whySuitable: string;
}

export interface CareerRecommendResponse {
  safeCareer: CareerOption;
  balancedCareer: CareerOption;
  dreamCareer: CareerOption;
}

export interface CareerSimulateBody {
  career: string;
  userProfile?: string;
}

export interface CareerSimulateResponse {
  dayInLife: string;
  salaryGrowth: string;
  realityCheck: string;
}

export interface SurvivalAnalysisBody {
  career: string;
  lifestyle?: string;
  stressLevel?: string;
  workHoursPreference?: string;
}

export interface SurvivalAnalysisResponse {
  survivalScore: number;
  lifestyleMatch: string;
  stressCompatibility: string;
  workHoursAnalysis: string;
  feedback: string;
  alternatives: string[];
}

export interface SkillGapBody {
  career: string;
  currentSkills: string;
}

export interface SkillItem {
  name: string;
  have: boolean;
  importance: string;
}

export interface SkillGapResponse {
  skills: SkillItem[];
  progressPercent: number;
  summary: string;
}

export interface JobRolesBody {
  career: string;
}

export interface JobRole {
  level: string;
  title: string;
  salary: string;
  description: string;
  requirements: string;
}

export interface JobRolesResponse {
  roles: JobRole[];
}

export interface DailyPlanBody {
  career: string;
  currentLevel?: string;
}

export interface DailyTask {
  task: string;
  duration: string;
  category: string;
}

export interface DailyPlanResponse {
  tasks: DailyTask[];
  motivation: string;
}

export interface CareerExplainBody {
  career: string;
  userProfile: string;
}

export interface CareerExplainResponse {
  whySuitable: string;
  whatToExpect: string;
}

export interface RoadmapStep {
  step: number;
  title: string;
  description: string;
  timeline: string;
  resources: string;
}

export interface RoadmapGenerateBody {
  career: string;
  currentLevel?: string;
  timeframe?: string;
}

export interface RoadmapGenerateResponse {
  steps: RoadmapStep[];
  totalTimeline: string;
  summary: string;
}

export interface ResumeAnalyzeBody {
  resumeText: string;
  targetCareer?: string;
}

export interface ResumeAnalyzeResponse {
  skillsFound: string[];
  missingSkills: string[];
  suggestedCareer: string;
  improvements: string[];
  overallScore: number;
  summary: string;
}

export interface Course {
  name: string;
  provider: string;
  url: string;
  isPaid: boolean;
  duration: string;
  level: string;
  description: string;
}

export interface CoursesListBody {
  career: string;
  level?: string;
}

export interface CoursesListResponse {
  freeCourses: Course[];
  paidCourses: Course[];
}

export interface LearningCenter {
  name: string;
  location: string;
  type: string;
  isPaid: boolean;
  description: string;
  contact: string;
}

export interface LearningCentersBody {
  location: string;
  career?: string;
}

export interface LearningCentersResponse {
  centers: LearningCenter[];
}