# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Career Navigator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /career/recommend:
    post:
      operationId: recommendCareers
      tags: [career]
      summary: Get AI career recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerRecommendBody"
      responses:
        "200":
          description: Career recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerRecommendResponse"
  /career/simulate:
    post:
      operationId: simulateCareer
      tags: [career]
      summary: Simulate a career day in life
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerSimulateBody"
      responses:
        "200":
          description: Career simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerSimulateResponse"
  /career/survival:
    post:
      operationId: survivalAnalysis
      tags: [career]
      summary: Get survival analysis for a career
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurvivalAnalysisBody"
      responses:
        "200":
          description: Survival analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurvivalAnalysisResponse"
  /career/skill-gap:
    post:
      operationId: analyzeSkillGap
      tags: [career]
      summary: Analyze skill gap for a career
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillGapBody"
      responses:
        "200":
          description: Skill gap analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGapResponse"
  /career/job-roles:
    post:
      operationId: getJobRoles
      tags: [career]
      summary: Get job roles for a career
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobRolesBody"
      responses:
        "200":
          description: Job roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRolesResponse"
  /career/daily-plan:
    post:
      operationId: getDailyPlan
      tags: [career]
      summary: Get daily action plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyPlanBody"
      responses:
        "200":
          description: Daily action plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlanResponse"
  /career/explain:
    post:
      operationId: explainCareer
      tags: [career]
      summary: Explain why a career suits the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerExplainBody"
      responses:
        "200":
          description: Career explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerExplainResponse"
...
  /courses/list:
    post:
      operationId: listCourses
      tags: [courses]
      summary: List courses for a career
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoursesListBody"
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursesListResponse"
  /courses/centers:
    post:
      operationId: listLearningCenters
      tags: [courses]
      summary: List nearby learning centers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningCentersBody"
      responses:
        "200":
          description: Learning centers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningCentersResponse"
components:
  schemas:
...
    CareerRecommendBody:
      type: object
      properties:
        interests:
          type: string
        skills:
          type: string
        skillLevel:
          type: string
        goals:
          type: string
        workLifePreference:
          type: string
        field:
          type: string
      required:
        - interests
        - skills
        - goals
    CareerOption:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        matchPercent:
          type: integer
        salary:
          type: string
        demand:
          type: string
        whySuitable:
          type: string
      required:
        - name
        - type
        - matchPercent
        - salary
        - demand
        - whySuitable
    CareerRecommendResponse:
      type: object
      properties:
        safeCareer:
          $ref: "#/components/schemas/CareerOption"
        balancedCareer:
          $ref: "#/components/schemas/CareerOption"
        dreamCareer:
          $ref: "#/components/schemas/CareerOption"
      required:
        - safeCareer
        - balancedCareer
        - dreamCareer
    CareerSimulateBody:
      type: object
      properties:
        career:
          type: string
        userProfile:
          type: string
      required:
        - career
    CareerSimulateResponse:
      type: object
      properties:
        dayInLife:
          type: string
        salaryGrowth:
          type: string
        realityCheck:
          type: string
      required:
        - dayInLife
        - salaryGrowth
        - realityCheck
    SurvivalAnalysisBody:
      type: object
      properties:
        career:
          type: string
        lifestyle:
          type: string
        stressLevel:
          type: string
        workHoursPreference:
          type: string
      required:
        - career
    SurvivalAnalysisResponse:
      type: object
      properties:
        survivalScore:
          type: integer
        lifestyleMatch:
          type: string
        stressCompatibility:
          type: string
        workHoursAnalysis:
          type: string
        feedback:
          type: string
        alternatives:
          type: array
          items:
            type: string
      required:
        - survivalScore
        - lifestyleMatch
        - stressCompatibility
        - workHoursAnalysis
        - feedback
        - alternatives
    SkillGapBody:
      type: object
      properties:
        career:
          type: string
        currentSkills:
          type: string
      required:
        - career
        - currentSkills
    SkillItem:
      type: object
      properties:
        name:
          type: string
        have:
          type: boolean
        importance:
          type: string
      required:
        - name
        - have
        - importance
    SkillGapResponse:
      type: object
      properties:
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillItem"
        progressPercent:
          type: integer
        summary:
          type: string
      required:
        - skills
        - progressPercent
        - summary
    JobRolesBody:
      type: object
      properties:
        career:
          type: string
      required:
        - career
    JobRole:
      type: object
      properties:
        level:
          type: string
        title:
          type: string
        salary:
          type: string
        description:
          type: string
        requirements:
          type: string
      required:
        - level
        - title
        - salary
        - description
        - requirements
    JobRolesResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/JobRole"
      required:
        - roles
    DailyPlanBody:
      type: object
      properties:
        career:
          type: string
        currentLevel:
          type: string
      required:
        - career
    DailyTask:
      type: object
      properties:
        task:
          type: string
        duration:
          type: string
        category:
          type: string
      required:
        - task
        - duration
        - category
    DailyPlanResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/DailyTask"
        motivation:
          type: string
      required:
        - tasks
        - motivation
    CareerExplainBody:
      type: object
      properties:
        career:
          type: string
        userProfile:
          type: string
      required:
        - career
        - userProfile
    CareerExplainResponse:
      type: object
      properties:
        whySuitable:
          type: string
        whatToExpect:
          type: string
      required:
        - whySuitable
        - whatToExpect
...
    Course:
      type: object
      properties:
        name:
          type: string
        provider:
          type: string
        url:
          type: string
        isPaid:
          type: boolean
        duration:
          type: string
        level:
          type: string
        description:
          type: string
      required:
        - name
        - provider
        - url
        - isPaid
        - duration
        - level
        - description
    CoursesListBody:
      type: object
      properties:
        career:
          type: string
        level:
          type: string
      required:
        - career
    CoursesListResponse:
      type: object
      properties:
        freeCourses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        paidCourses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
      required:
        - freeCourses
        - paidCourses
    LearningCenter:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        type:
          type: string
        isPaid:
          type: boolean
        description:
          type: string
        contact:
          type: string
      required:
        - name
        - location
        - type
        - isPaid
        - description
        - contact
    LearningCentersBody:
      type: object
      properties:
        location:
          type: string
        career:
          type: string
      required:
        - location
    LearningCentersResponse:
      type: object
      properties:
        centers:
          type: array
          items:
            $ref: "#/components/schemas/LearningCenter"
      required:
        - centers