# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Navigator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Get AI career recommendations
 */
export const RecommendCareersBody = zod.object({
  interests: zod.string(),
  skills: zod.string(),
  skillLevel: zod.string().optional(),
  goals: zod.string(),
  workLifePreference: zod.string().optional(),
  field: zod.string().optional(),
});

export const RecommendCareersResponse = zod.object({
  safeCareer: zod.object({
    name: zod.string(),
    type: zod.string(),
    matchPercent: zod.number(),
    salary: zod.string(),
    demand: zod.string(),
    whySuitable: zod.string(),
  }),
  balancedCareer: zod.object({
    name: zod.string(),
    type: zod.string(),
    matchPercent: zod.number(),
    salary: zod.string(),
    demand: zod.string(),
    whySuitable: zod.string(),
  }),
  dreamCareer: zod.object({
    name: zod.string(),
    type: zod.string(),
    matchPercent: zod.number(),
    salary: zod.string(),
    demand: zod.string(),
    whySuitable: zod.string(),
  }),
});

/**
 * @summary Simulate a career day in life
 */
export const SimulateCareerBody = zod.object({
  career: zod.string(),
  userProfile: zod.string().optional(),
});

export const SimulateCareerResponse = zod.object({
  dayInLife: zod.string(),
  salaryGrowth: zod.string(),
  realityCheck: zod.string(),
});

/**
 * @summary Get survival analysis for a career
 */
export const SurvivalAnalysisBody = zod.object({
  career: zod.string(),
  lifestyle: zod.string().optional(),
  stressLevel: zod.string().optional(),
  workHoursPreference: zod.string().optional(),
});

export const SurvivalAnalysisResponse = zod.object({
  survivalScore: zod.number(),
  lifestyleMatch: zod.string(),
  stressCompatibility: zod.string(),
  workHoursAnalysis: zod.string(),
  feedback: zod.string(),
  alternatives: zod.array(zod.string()),
});

/**
 * @summary Analyze skill gap for a career
 */
export const AnalyzeSkillGapBody = zod.object({
  career: zod.string(),
  currentSkills: zod.string(),
});

export const AnalyzeSkillGapResponse = zod.object({
  skills: zod.array(
    zod.object({
      name: zod.string(),
      have: zod.boolean(),
      importance: zod.string(),
    }),
  ),
  progressPercent: zod.number(),
  summary: zod.string(),
});

/**
 * @summary Get job roles for a career
 */
export const GetJobRolesBody = zod.object({
  career: zod.string(),
});

export const GetJobRolesResponse = zod.object({
  roles: zod.array(
    zod.object({
      level: zod.string(),
      title: zod.string(),
      salary: zod.string(),
      description: zod.string(),
      requirements: zod.string(),
    }),
  ),
});

/**
 * @summary Get daily action plan
 */
export const GetDailyPlanBody = zod.object({
  career: zod.string(),
  currentLevel: zod.string().optional(),
});

export const GetDailyPlanResponse = zod.object({
  tasks: zod.array(
    zod.object({
      task: zod.string(),
      duration: zod.string(),
      category: zod.string(),
    }),
  ),
  motivation: zod.string(),
});

/**
 * @summary Explain why a career suits the user
 */
export const ExplainCareerBody = zod.object({
  career: zod.string(),
  userProfile: zod.string(),
});

export const ExplainCareerResponse = zod.object({
  whySuitable: zod.string(),
  whatToExpect: zod.string(),
});

/**
 * @summary Generate a personalized career roadmap
 */
export const GenerateRoadmapBody = zod.object({
  career: zod.string(),
  currentLevel: zod.string().optional(),
  timeframe: zod.string().optional(),
});

export const GenerateRoadmapResponse = zod.object({
  steps: zod.array(
    zod.object({
      step: zod.number(),
      title: zod.string(),
      description: zod.string(),
      timeline: zod.string(),
      resources: zod.string(),
    }),
  ),
  totalTimeline: zod.string(),
  summary: zod.string(),
});

/**
 * @summary Analyze uploaded resume (PDF or text)
 */
export const AnalyzeResumeBody = zod.object({
  resumeText: zod.string(),
  targetCareer: zod.string().optional(),
});

export const AnalyzeResumeResponse = zod.object({
  skillsFound: zod.array(zod.string()),
  missingSkills: zod.array(zod.string()),
  suggestedCareer: zod.string(),
  improvements: zod.array(zod.string()),
  overallScore: zod.number(),
  summary: zod.string(),
});

/**
 * @summary List courses for a career
 */
export const ListCoursesBody = zod.object({
  career: zod.string(),
  level: zod.string().optional(),
});

export const ListCoursesResponse = zod.object({
  freeCourses: zod.array(
    zod.object({
      name: zod.string(),
      provider: zod.string(),
      url: zod.string(),
      isPaid: zod.boolean(),
      duration: zod.string(),
      level: zod.string(),
      description: zod.string(),
    }),
  ),
  paidCourses: zod.array(
    zod.object({
      name: zod.string(),
      provider: zod.string(),
      url: zod.string(),
      isPaid: zod.boolean(),
      duration: zod.string(),
      level: zod.string(),
      description: zod.string(),
    }),
  ),
});

/**
 * @summary List nearby learning centers
 */
export const ListLearningCentersBody = zod.object({
  location: zod.string(),
  career: zod.string().optional(),
});

export const ListLearningCentersResponse = zod.object({
  centers: zod.array(
    zod.object({
      name: zod.string(),
      location: zod.string(),
      type: zod.string(),
      isPaid: zod.boolean(),
      description: zod.string(),
      contact: zod.string(),
    }),
  ),
});