# lib/api-zod/src/generated/types/careerSimulateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Career Navigator API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerSimulateBody {
  career: string;
  userProfile?: string;
}